Lock utils moved to a separate file
